separated aws apis
