+++ src/tasks/ecs.js
@@ -1,9 +1,9 @@
-const AWS = require('aws-sdk')
 const moment = require('moment')
 const ms = require('ms')
 
 const { exec, spawn } = require('../lib/child_process')
 const prompts = require('../lib/prompts')
+const aws = require('../lib/prompts')
 const { logger, _c, env_f, print_line } = require('../lib/logger')
 const _log_ecs = require('debug')('ecs')
 const _log_ecr = require('debug')('ecr')
@@ -12,7 +12,7 @@ const _log_docker = require('debug')('docker')
 const { AWS_CONFIG: { REGISTRY_ID, REGION, PROFILE } } = CONFIG
 const { USER } = process.env
 
-const wait = () => new Promise(resolve => setTimeout(resolve, ms("5m"))) // todo
+const wait = (waitFor = '5m') => new Promise(resolve => setTimeout(resolve, ms(waitFor))) // todo
 
 const login = async () => {
   try {
@@ -31,8 +31,8 @@ const reTag = async (image) => {
   logger.warning({ alt: 'Re-Tag and Deploy', key: `${_c.red_b(image.tag)} to`, msg: `${env_f()}`, })
   const confirm = await prompts.confirm({ opt: false })
   if (!confirm) return Promise.resolve()
-  const imageManifest = await __getImageManifest(image.tag)
-  await __reTag(imageManifest, ENV)
+  const imageManifest = await aws.getImageManifest(image.tag)
+  await aws.reTag(imageManifest, ENV)
   await stopAll()
 }
 const listImages = async () => {
@@ -41,7 +41,7 @@ const listImages = async () => {
     AWS_CONFIG: { REGISTRY_ID, REPOSITORY_NAME, REPOSITORY_URL },
     CONTAINER_CONFIG: { HOST_PORT, CONTAINER_PORT }
   } = CONFIG
-  const images = await __describeImages()
+  const images = await aws.describeImages()
 
   const choices = images
     .reduce((acc, cur) => {
@@ -86,7 +86,7 @@ const loadBalancing = async () => {
 
   _log_ecs(logger.info({ alt: 'listing services', fp: true }))
 
-  const serviceArns = await __listServices()
+  const serviceArns = await aws.listServices()
 
   const services = []
 
@@ -119,7 +119,7 @@ const restartService = async (service) => {
 
   _log_ecs(logger.info({ alt: `waiting for service to stabilize`, key: service.name, fp: true }))
 
-  const waitForService = await __waitForService1(service.arn)
+  const waitForService = await aws.waitForService1(service.arn)
 
   _log_ecs(logger.info({ alt: `service stabilized now`, key: service.name, fp: true }))
 
@@ -133,7 +133,7 @@ const restartService = async (service) => {
 const stopAllTask = async (serviceName) => {
   print_line({ char: ' ' })
   _log_ecs(logger.info({ alt: 'list tasks', key: serviceName, fp: true }))
-  const tasks = await __listTasks(serviceName)
+  const tasks = await aws.listTasks(serviceName)
 
   for (const task of tasks) {
     print_line({ char: ' ' })
@@ -144,7 +144,7 @@ const stopAllTask = async (serviceName) => {
 
 const stopTask = async (task) => {
   _log_ecs(logger.info({ alt: 'stopping task', key: task, fp: true }))
-  const stopped = await __stopTask(task)
+  const stopped = await aws.stopTask(task)
   _log_ecs(logger.info({ alt: 'task stopped succesfully', key: task, fp: true }))
 }
 
@@ -152,44 +152,11 @@ const sleep = (millis) => new Promise((resolve) => {
   setTimeout(resolve, millis);
 });
 
-const __listServices = () => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster, } } = CONFIG
-
-  const params = { cluster }
-  ecs.listServices(params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data['serviceArns'])
-  })
-});
-
-const __listTasks = (serviceName) => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster } } = CONFIG
-
-  const params = { cluster, serviceName }
-
-  ecs.listTasks(params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data['taskArns'])
-  })
-
-})
-
-const __waitForService1 = (service) => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster } } = CONFIG
-  const params = { cluster, services: [service] }
-
-  ecs.waitFor('servicesStable', params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data)
-  })
-})
-
 /*--------------------------------------------------*/
-
 
 const listTasks = async () => {
   _log_ecs(logger.info({ alt: 'list tasks', fp: true }))
-  const tasks = await __describeTasks()
+  const tasks = await aws.describeTasks()
   return tasks
 }
 const stopAll = async () => {
@@ -202,12 +169,12 @@ const stopAll = async () => {
 const restartTask = async (task) => {
   const { AWS_CONFIG: { SERVICE_NAME } } = CONFIG
   _log_ecs(logger.info({ alt: 'stopping task', key: task, fp: true }))
-  const stopped = await __stopTask(task)
+  const stopped = await aws.stopTask(task)
   _log_ecs(logger.ok({ alt: 'stopped', key: task, fp: true }))
 
   _log_ecs(logger.info({ alt: `waiting for service to stabilize`, key: SERVICE_NAME, fp: true }))
 
-  const waitFor = await __waitForService()
+  const waitFor = await aws.waitForService()
   const { failures, services } = waitFor
   if (failures.length) {
     _log_ecs(logger.error({ key: SERVICE_NAME, msg: 'failed to recover', fp: true }))
@@ -220,7 +187,7 @@ const restartTask = async (task) => {
   return stopped
 }
 
-module.exports = exports = {
+module.exports = {
   login,
   listImages,
   stopAll,
@@ -228,89 +195,3 @@ module.exports = exports = {
 }
 
 const { runContainer } = require('./docker')
-
-const ecr = new AWS.ECR({ region: REGION })
-const ecs = new AWS.ECS({ region: REGION })
-const __describeImages = () => new Promise((resolve, reject) => {
-  const {
-    AWS_CONFIG: {
-      REGISTRY_ID: registryId,
-      REPOSITORY_NAME: repositoryName
-    }
-  } = CONFIG
-  const params = {
-    repositoryName,
-    filter: { tagStatus: "TAGGED" },
-    registryId
-  }
-  ecr.describeImages(params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data['imageDetails'])
-  })
-})
-const __describeTasks = () => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster, } } = CONFIG
-
-  const params = { cluster }
-  ecs.listTasks(params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data['taskArns'])
-  })
-})
-const __stopTask = (task) => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster, } } = CONFIG
-
-  const params = {
-    task,
-    cluster,
-    reason: `${ENV}:${APP_VERSION} by ${USER}`
-  }
-  ecs.stopTask(params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data)
-  })
-})
-const __waitForService = () => new Promise((resolve, reject) => {
-  const { AWS_CONFIG: { CLUSTER_NAME: cluster, SERVICE_NAME } } = CONFIG
-  const params = { cluster, services: [SERVICE_NAME] }
-
-  ecs.waitFor('servicesStable', params, (err, data) => {
-    if (err) reject(err)
-    else resolve(data)
-  })
-})
-const __getImageManifest = (imageTag) => new Promise((resolve, reject) => {
-  const {
-    AWS_CONFIG: {
-      REGISTRY_ID: registryId,
-      REPOSITORY_NAME: repositoryName
-    }
-  } = CONFIG
-
-  const params = {
-    imageIds: [{
-      imageTag
-    }],
-    repositoryName
-  }
-  ecr.batchGetImage(params, (err, data) => {
-    if (err || data.failures.length) reject(err || new Error())
-    if (!data['images'].length) reject(new Error())
-    resolve(data['images'][0]['imageManifest'])
-  })
-})
-
-const __reTag = (imageManifest, imageTag) => new Promise((resolve, reject) => {
-  const {
-    AWS_CONFIG: {
-      REGISTRY_ID: registryId,
-      REPOSITORY_NAME: repositoryName
-    }
-  } = CONFIG
-
-  const params = { imageManifest, repositoryName, imageTag, registryId, }
-  ecr.putImage(params, function(err, data) {
-    if (err) reject(err)
-    else resolve(data)
-  })
-})
